small refactoring, changes components folders. created a back button
