+++ frontend/src/App/styles.tsx
@@ -15,18 +15,7 @@ const useStyles = makeStyles((theme: Theme) =>
     appbar: {
       'margin-bottom': '40px'
     },
-    progressBox: {
 
-      display: 'flex',
-      flexFlow: "flex-direction",
-      paddingTop: '10px'
-    },
-    progress: {
-      width: '200px',
-      marginTop: '10px',
-      marginLeft: '5px',
-      marginRight: '5px',
-    },
     menuLeft: {
       display: 'flex',
       flexFlow: "flex-direction",
@@ -35,14 +24,6 @@ const useStyles = makeStyles((theme: Theme) =>
       height: '30px',
       marginTop: '9px'
     },
-    progressContainer: {
-      'display': 'flex'
-    },
-
-    resetButton: {
-      color: '#FFF',
-      marginLeft: '5px'
-    }
 
   }),
 );
